refactor: Project setup
