+++ drizzle.config.ts
@@ -1,5 +1,6 @@
+import { env } from "@/env";
 import { defineConfig } from "drizzle-kit";
-import { env } from "./env";
+
 export default defineConfig({
   dialect: "sqlite",
   schema: "./src/db/schema.ts",
