+++ package.json
@@ -3,36 +3,38 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "next dev",
     "build": "next build",
-    "start": "next start",
-    "lint": "next lint",
+    "dev": "next dev",
+    "format": "prettier --write .",
     "generate": "drizzle-kit generate",
-    "migrate": "drizzle-kit migrate"
+    "lint": "next lint",
+    "migrate": "drizzle-kit migrate",
+    "start": "next start"
   },
   "dependencies": {
     "@apollo/server": "^4.10.4",
     "@as-integrations/next": "^3.0.0",
     "@libsql/client": "^0.7.0",
     "@t3-oss/env-nextjs": "^0.10.1",
-    "@types/bun": "^1.1.6",
     "drizzle-orm": "^0.32.0",
     "graphql": "^16.9.0",
     "next": "14.2.5",
-    "react": "^18",
-    "react-dom": "^18",
-    "sqlite": "^5.1.1",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "zod": "^3.23.8"
   },
   "devDependencies": {
-    "@types/node": "^20",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
+    "@ianvs/prettier-plugin-sort-imports": "^4.3.1",
+    "@types/node": "^20.14.11",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
     "drizzle-kit": "^0.23.0",
     "eslint": "^8",
     "eslint-config-next": "14.2.5",
-    "postcss": "^8",
+    "postcss": "^8.4.39",
+    "prettier": "^3.3.3",
+    "prettier-plugin-tailwindcss": "^0.6.5",
     "tailwindcss": "^3.4.1",
-    "typescript": "^5"
+    "typescript": "^5.5.3"
   }
 }
