+++ src/app/api/graphql/schema.graphql
@@ -1,0 +1,3 @@
+type Query {
+  hello: String
+}
