--- src/app/api/graphql/schema.ts
@@ -1,5 +1,0 @@
-export const typeDefs = `#graphql
-    type Query {
-        hello: String
-    }
-`;
