+++ src/lib/db/db.ts
@@ -1,6 +1,7 @@
+import { env } from "@/env";
 import { createClient } from "@libsql/client";
 
 export const turso = createClient({
-  url: Bun.env.TURSO_DATABASE_URL,
-  authToken: Bun.env.TURSO_AUTH_TOKEN,
+  url: env.TURSO_DATABASE_URL,
+  authToken: env.TURSO_AUTH_TOKEN,
 });
