+++ src/lib/db/schema.ts
@@ -1,4 +1,5 @@
-import { serial, text, timestamp, pgTable } from "drizzle-orm/pg-core";
+import { pgTable, serial, text, timestamp } from "drizzle-orm/pg-core";
+
 export const user = pgTable("user", {
   id: serial("id"),
   name: text("name"),
